+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Change `FactoryBot/CreateList` so that it considers `times.map`. ([@r7kamura])
 - Add `FactoryBot/RedundantFactoryOption` cop. ([@r7kamura])
 - Add `ExplicitOnly` configuration option to `FactoryBot/ConsistentParenthesesStyle`, `FactoryBot/CreateList` and `FactoryBot/FactoryNameStyle`. ([@ydah])
+- Change `FactoryBot/CreateList` so that it checks same factory calls in an Array. ([@r7kamura])
 
 ## 2.22.0 (2023-05-04)